Extracted view_helpers.py
