+++ corehq/messaging/scheduling/tests/test_bulk_conditional_alerts.py
@@ -29,7 +29,7 @@ from corehq.apps.users.models import CommCareUser
 from corehq.messaging.scheduling.scheduling_partitioned.dbaccessors import delete_case_schedule_instance
 from corehq.messaging.scheduling.scheduling_partitioned.models import CaseTimedScheduleInstance
 from corehq.messaging.scheduling.tests.util import delete_timed_schedules
-from corehq.messaging.scheduling.views import get_conditional_alert_rows, upload_conditional_alert_workbook
+from corehq.messaging.scheduling.view_helpers import get_conditional_alert_rows, upload_conditional_alert_workbook
 from corehq.sql_db.util import run_query_across_partitioned_databases
 from corehq.util.workbook_json.excel import get_workbook
