+++ corehq/messaging/scheduling/views.py
@@ -52,6 +52,11 @@ from corehq.messaging.scheduling.scheduling_partitioned.dbaccessors import (
 from corehq.messaging.scheduling.tasks import refresh_alert_schedule_instances, refresh_timed_schedule_instances
 from corehq.messaging.tasks import initiate_messaging_rule_run
 from corehq.messaging.util import MessagingRuleProgressHelper
+from corehq.messaging.scheduling.view_helpers import (
+    get_conditional_alert_rows,
+    get_conditional_alerts_queryset_by_domain,
+    upload_conditional_alert_workbook,
+)
 from corehq.const import SERVER_DATETIME_FORMAT
 from corehq.util.timezones.conversions import ServerTime
 from corehq.util.timezones.utils import get_timezone_for_user
@@ -561,21 +566,8 @@ class ConditionalAlertBaseView(BaseMessagingSectionView):
     def dispatch(self, *args, **kwargs):
         return super(ConditionalAlertBaseView, self).dispatch(*args, **kwargs)
 
-    @staticmethod
-    def get_conditional_alerts_queryset_by_domain(domain, query_string=''):
-        query = (
-            AutomaticUpdateRule
-            .objects
-            .filter(domain=domain, workflow=AutomaticUpdateRule.WORKFLOW_SCHEDULING, deleted=False)
-        )
-        if query_string:
-            query = query.filter(name__icontains=query_string)
-        query = query.order_by('case_type', 'name', 'id')
-        return query
-
     def get_conditional_alerts_queryset(self, query_string=''):
-        return ConditionalAlertBaseView.get_conditional_alerts_queryset_by_domain(self.domain,
-                                                                                  query_string=query_string)
+        return get_conditional_alerts_queryset_by_domain(self.domain, query_string=query_string)
 
 
 class ConditionalAlertListView(ConditionalAlertBaseView):
@@ -1046,26 +1038,6 @@ class DownloadConditionalAlertView(ConditionalAlertBaseView):
         return export_response(temp, Format.XLS_2007, filename)
 
 
-# TODO: move to a view_helpers.py file
-def get_conditional_alert_rows(domain, langs):
-    translated_rows = []
-    untranslated_rows = []
-
-    for rule in ConditionalAlertBaseView.get_conditional_alerts_queryset_by_domain(domain):
-        if not isinstance(_get_rule_content(rule), SMSContent):
-            continue
-        common_columns = [rule.pk, rule.name, rule.case_type]
-        message = ContentForm.compute_initial(
-            rule.get_messaging_rule_schedule().memoized_events[0].content
-        ).get('message', {})
-        if '*' in message or len(message) == 0:
-            untranslated_rows.append(common_columns + [message.get('*', '')])
-        else:
-            translated_rows.append(common_columns + [message.get(lang, '') for lang in langs])
-
-    return (translated_rows, untranslated_rows)
-
-
 class UploadConditionalAlertView(BaseMessagingSectionView):
     urlname = 'upload_conditional_alert'
     page_title = ugettext_lazy("Upload Conditional Alerts")
@@ -1112,123 +1084,3 @@ class UploadConditionalAlertView(BaseMessagingSectionView):
             msg[0](request, msg[1])
 
         return self.get(request, *args, **kwargs)
-
-
-# TODO: move to a view_helpers.py file
-def upload_conditional_alert_workbook(domain, langs, workbook):
-    translated_uploader = TranslatedConditionalAlertUploader(domain, langs)
-    untranslated_uploader = UntranslatedConditionalAlertUploader(domain, langs)
-    return translated_uploader.upload(workbook) + untranslated_uploader.upload(workbook)
-
-
-# TODO: move these to a view_helpers.py file
-class ConditionalAlertUploader(object):
-    sheet_name = None
-
-    def __init__(self, domain, langs):
-        super(ConditionalAlertUploader, self).__init__()
-        self.domain = domain
-        self.langs = langs
-        self.msgs = []
-
-    def applies_to_rule(self, rule):
-        return True
-
-    def rule_message(self, rule):
-        return ContentForm.compute_initial(
-            rule.get_messaging_rule_schedule().memoized_events[0].content
-        ).get('message', {})
-
-    def upload(self, workbook):
-        self.msgs = []
-        success_count = 0
-        rows = workbook.get_worksheet(title=self.sheet_name)
-
-        for index, row in enumerate(rows, start=2):    # one-indexed, plus header row
-            rule = None
-            try:
-                rule = AutomaticUpdateRule.objects.get(
-                    pk=row['id'],
-                    domain=self.domain,
-                    workflow=AutomaticUpdateRule.WORKFLOW_SCHEDULING,
-                    deleted=False,
-                )
-            except AutomaticUpdateRule.DoesNotExist:
-                self.msgs.append((messages.error,
-                                 _("""Could not find rule for row {index} in '{sheet_name}' sheet, """
-                                   """with id {id}""").format(index=index,
-                                                              id=row['id'],
-                                                              sheet_name=self.sheet_name)))
-
-            if rule:
-                if not isinstance(_get_rule_content(rule), SMSContent):
-                    self.msgs.append((messages.error, _("Row {index} in '{sheet_name}' sheet, with rule id {id}, "
-                                      "does not use SMS content.").format(index=index, id=row['id'],
-                                                                          sheet_name=self.sheet_name)))
-                elif self.applies_to_rule(rule):
-                    dirty = self.update_rule(rule, row)
-                    if dirty:
-                        rule.save()
-                        success_count += 1
-                else:
-                    self.msgs.append((messages.error, _("Rule in row {index} with id {id} does not belong in " \
-                        "'{sheet_name}' sheet.".format(index=index, id=row['id'], sheet_name=self.sheet_name))))
-
-        self.msgs.append((messages.success, _("Updated {count} rule(s) in '{sheet_name}' sheet").format(
-            count=success_count, sheet_name=self.sheet_name)))
-
-        return self.msgs
-
-    def update_rule(self, rule, row):
-        dirty = False
-        if rule.name != row['name']:
-            dirty = True
-            rule.name = row['name']
-        if rule.case_type != row['case_type']:
-            dirty = True
-            rule.case_type = row['case_type']
-        return dirty
-
-    def update_rule_message(self, rule, message):
-        # TODO
-        pass
-
-
-class TranslatedConditionalAlertUploader(ConditionalAlertUploader):
-    sheet_name = 'translated'
-
-    def applies_to_rule(self, rule):
-        message = self.rule_message(rule)
-        return message and '*' not in message
-
-    def update_rule(self, rule, row):
-        dirty = super(TranslatedConditionalAlertUploader, self).update_rule(rule, row)
-        message = self.rule_message(rule)
-        message_dirty = False
-        for lang in self.langs:
-            if message.get(lang, '') != row['message_' + lang]:
-                message.update({lang: row['message_' + lang]})
-                message_dirty = True
-        if message_dirty:
-            self.update_rule_message(rule, message)
-        return dirty or message_dirty
-
-
-class UntranslatedConditionalAlertUploader(ConditionalAlertUploader):
-    sheet_name = 'not translated'
-
-    def applies_to_rule(self, rule):
-        message = self.rule_message(rule)
-        return not message or '*' in message
-
-    def update_rule(self, rule, row):
-        dirty = super(UntranslatedConditionalAlertUploader, self).update_rule(rule, row)
-        message = self.rule_message(rule)
-        if message.get('*', '') != row['message']:
-            self.update_rule_message(rule, message)
-            dirty = True
-        return dirty
-
-
-def _get_rule_content(rule):
-    return rule.get_messaging_rule_schedule().memoized_events[0].content
